ouu that's a good commit

+++ PC_Tracker/APU.SDL.OLD.DONT.TOUCH/gme_vspc/gme/gme/Spc_Dsp_Register_Map_Interface.h
@@ -124,7 +124,7 @@ enum {
 };
 
   struct reg_lut_t {
-  uint8_t addr;
+  int addr;
   const char *name;
   //char *altname;
 };
